refactor(credit): modularize market_data_engine → market_data/ package (Wave 1 PR #7)
Restructure market_data_engine.py (1031 LOC) into edgar-style market_data/ package:
- models.py: FRED_SERIES_REGISTRY, CASE_SHILLER_METRO_MAP, GEOGRAPHY_TO_METRO, constants
- fred_client.py: _fetch_fred_series, _fetch_latest_strict, _latest_value
- computed_fields.py: yield curve 2s10s, CPI YoY, GDP growth
- regional.py: resolve_metro_key, fetch_regional_case_shiller
- stress.py: compute_macro_stress_severity, compute_macro_stress_flag
- snapshot.py: legacy v1 + expanded v2 snapshot builders
- service.py: get_macro_snapshot (daily-cached orchestrator)
- __init__.py: PEP 562 lazy imports (httpx heavy dep), __all__ with 6 exports

Error contract: never-raises (orchestration engine called during deep review).
structlog replaces stdlib logging.
