+++ backend/tests/test_vertical_engines.py
@@ -40,7 +40,7 @@ class TestCreditModuleStructure:
         "vertical_engines.credit.deep_review",
         "vertical_engines.credit.critic",
         "vertical_engines.credit.quant",
-        "vertical_engines.credit.market_data_engine",
+        "vertical_engines.credit.market_data",
         "vertical_engines.credit.sponsor",
         "vertical_engines.credit.memo_chapter_engine",
         "vertical_engines.credit.memo_book_generator",
