+++ backend/vertical_engines/credit/market_data/__init__.py
@@ -1,0 +1,56 @@
+"""Market Data Engine v2 — deterministic FRED macro overlay for Deep Review v4.
+
+Public API:
+    get_macro_snapshot()              — daily-cached macro snapshot (orchestrator)
+    compute_macro_stress_severity()   — graded stress assessment
+    compute_macro_stress_flag()       — legacy bool stress flag
+    resolve_metro_key()               — geography → Case-Shiller metro key
+    fetch_regional_case_shiller()     — regional HPI fetch
+
+Error contract: never-raises (orchestration engine called during deep review).
+Returns snapshot with stress severity embedded.
+"""
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from typing import Any
+
+
+def __getattr__(name: str) -> Any:
+    if name == "get_macro_snapshot":
+        from vertical_engines.credit.market_data.service import get_macro_snapshot
+
+        return get_macro_snapshot
+    if name == "compute_macro_stress_severity":
+        from vertical_engines.credit.market_data.stress import compute_macro_stress_severity
+
+        return compute_macro_stress_severity
+    if name == "compute_macro_stress_flag":
+        from vertical_engines.credit.market_data.stress import compute_macro_stress_flag
+
+        return compute_macro_stress_flag
+    if name == "resolve_metro_key":
+        from vertical_engines.credit.market_data.regional import resolve_metro_key
+
+        return resolve_metro_key
+    if name == "fetch_regional_case_shiller":
+        from vertical_engines.credit.market_data.regional import fetch_regional_case_shiller
+
+        return fetch_regional_case_shiller
+    if name == "FRED_SERIES_REGISTRY":
+        from vertical_engines.credit.market_data.models import FRED_SERIES_REGISTRY
+
+        return FRED_SERIES_REGISTRY
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+
+
+__all__ = [
+    "get_macro_snapshot",
+    "compute_macro_stress_severity",
+    "compute_macro_stress_flag",
+    "resolve_metro_key",
+    "fetch_regional_case_shiller",
+    "FRED_SERIES_REGISTRY",
+]
