+++ docs/plans/2026-03-15-refactor-credit-vertical-modular-alignment-wave1-plan.md
@@ -301,7 +301,7 @@ quant/
 
 ---
 
-### Phase 7: market_data/ (PR #7)
+### Phase 7: market_data/ (PR #7) ✅
 
 **Source:** `market_data_engine.py` (1031 LOC)
